fix(feishu): expand quoted merged-forward messages
Centralize merged-forward expansion in message retrieval so direct messages, replies, and thread reads share the same parsed content.
