+++ extensions/feishu/src/bot.helpers.test.ts
@@ -3,7 +3,8 @@ import { describe, expect, it } from "vitest";
 import type { ClawdbotConfig } from "../runtime-api.js";
 import { buildFeishuAgentBody } from "./bot-agent-body.js";
 import { buildBroadcastSessionKey, resolveBroadcastAgents } from "./bot-broadcast.js";
-import { parseMergeForwardContent, parseMessageContent } from "./bot-content.js";
+import { parseMessageContent } from "./bot-content.js";
+import { parseMergeForwardContent } from "./message-content.js";
 
 describe("buildFeishuAgentBody", () => {
   it("builds message id, speaker, quoted content, mention context, and permission notice in order", () => {
